@@ -15,85 +16,91 @@
     "tauri": "tauri"
   },
   "dependencies": {
-    "@hookform/resolvers": "^3.9.0",
-    "@radix-ui/react-alert-dialog": "^1.1.1",
-    "@radix-ui/react-checkbox": "^1.1.1",
+    "@hookform/resolvers": "^3.10.0",
+    "@radix-ui/react-alert-dialog": "^1.1.4",
+    "@radix-ui/react-checkbox": "^1.1.3",
     "@radix-ui/react-collapsible": "^1.1.1",
     "@radix-ui/react-dialog": "^1.1.2",
     "@radix-ui/react-dropdown-menu": "^2.1.1",
     "@radix-ui/react-label": "^2.1.0",
-    "@radix-ui/react-navigation-menu": "^1.2.0",
+    "@radix-ui/react-navigation-menu": "^1.2.3",
     "@radix-ui/react-popover": "^1.1.1",
     "@radix-ui/react-progress": "^1.1.0",
     "@radix-ui/react-radio-group": "^1.2.0",
     "@radix-ui/react-scroll-area": "^1.1.0",
-    "@radix-ui/react-select": "^2.1.1",
+    "@radix-ui/react-select": "^2.1.4",
     "@radix-ui/react-separator": "^1.1.0",
     "@radix-ui/react-slot": "^1.1.0",
     "@radix-ui/react-switch": "^1.1.0",
     "@radix-ui/react-toast": "^1.2.1",
-    "@radix-ui/react-tooltip": "^1.1.2",
+    "@radix-ui/react-tooltip": "^1.1.6",
     "@reduxjs/toolkit": "^2.2.6",
     "@tailwindcss/container-queries": "^0.1.1",
+    "@tailwindcss/postcss": "^4.0.0",
+    "@tailwindcss/vite": "^4.0.0",
+    "@tanstack/react-query": "^5.65.1",
+    "@tanstack/react-query-devtools": "^5.65.1",
     "@tanstack/react-table": "^8.17.3",
     "@tauri-apps/plugin-fs": "^2.0.0-rc.2",
-    "@tauri-apps/plugin-notification": "^2.0.0-beta.3",
+    "@tauri-apps/plugin-notification": "^2.2.1",
     "@tauri-apps/plugin-os": "^2.0.0-rc.0",
     "@tauri-apps/plugin-process": "^2.0.0-rc.0",
     "@tauri-apps/plugin-shell": ">=2.0.0-beta.0",
     "@tauri-apps/plugin-sql": "^2.0.0-rc.0",
-    "@tauri-apps/plugin-updater": "^2.0.0-rc.0",
-    "@types/animejs": "^3.1.12",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
+    "@tauri-apps/plugin-updater": "^2.4.0",
+    "@types/animejs": "^3.1.13",
     "animejs": "^3.2.2",
+    "axios": "^1.7.7",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "cmdk": "^1.0.0",
-    "date-fns": "^3.6.0",
-    "decimal.js": "^10.4.3",
+    "date-fns": "^4.1.0",
+    "decimal.js": "^10.5.0",
     "dexie": "^4.0.7",
     "dexie-react-hooks": "^1.1.7",
     "init": "^0.1.2",
     "lodash": "^4.17.21",
-    "lucide-react": "^0.390.0",
-    "react": "^18.2.0",
-    "react-day-picker": "^8.10.1",
-    "react-dom": "^18.2.0",
+    "lucide-react": "^0.474.0",
+    "react": "19.0.0",
+    "react-day-picker": "^9.5.1",
+    "react-dom": "19.0.0",
     "react-dropzone": "^14.2.3",
-    "react-hook-form": "^7.53.0",
+    "react-hook-form": "^7.54.2",
     "react-icons": "^5.2.1",
     "react-loader-spinner": "^6.1.6",
-    "react-redux": "^9.1.2",
+    "react-query": "^3.39.3",
+    "react-redux": "^9.2.0",
     "react-responsive": "^10.0.0",
-    "recharts": "^2.12.7",
+    "recharts": "^2.15.1",
     "tailwind-merge": "^2.3.0",
+    "tailwindcss": "^4.0.0",
     "tailwindcss-animate": "^1.0.7",
     "typescript": "^5.5.4",
-    "vaul": "^0.9.1",
-    "zod": "^3.23.8"
+    "vaul": "^1.1.2",
+    "zod": "^3.24.1"
   },
   "devDependencies": {
     "@faker-js/faker": "^9.0.0",
     "@jest/globals": "^29.7.0",
     "@playwright/test": "^1.47.1",
     "@tauri-apps/api": "^2.0.0-rc.0",
-    "@tauri-apps/cli": "^2.0.0-rc.3",
+    "@tauri-apps/cli": "^2.2.7",
     "@types/jest": "^29.5.13",
-    "@types/node": "^20.14.14",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
+    "@types/node": "^22.12.0",
+    "@types/react": "^19.0.8",
+    "@types/react-dom": "^19.0.3",
     "@vitejs/plugin-react": "^4.2.1",
     "autoprefixer": "^10.4.19",
     "chalk": "^5.3.0",
     "cli-progress": "^3.12.0",
     "cross-env": "^7.0.3",
-    "internal-ip": "^7.0.0",
+    "fake-indexeddb": "^6.0.0",
+    "internal-ip": "^8.0.0",
     "jest": "^29.7.0",
     "postcss": "^8.4.38",
-    "react-router-dom": "^6.23.1",
-    "tailwindcss": "^3.4.3",
+    "react-router-dom": "^7.1.3",
     "ts-jest": "^29.2.5",
-    "vite": "^5.4.3"
+    "tsconfig-paths": "^4.2.0",
+    "vite": "^6.0.11"
   }
 }
