@@ -1,25 +1,71 @@
-import Dexie, { BulkError, Collection, ModifyError } from "dexie";
-import { Birthday } from "./tables/birthday";
-import { getDefaultSettings, type Settings } from "./tables/setting";
-import { calcDaysUntilNextBirthday } from "@/lib/main_util";
+import Dexie from "dexie";
+import type { BulkError, Collection, ModifyError } from "dexie";
+
+import type { Birthday } from "./tables/birthday/birthdays.ts";
+import type { Settings } from "./tables/settings/settings.ts";
+
+// import calcDaysUntilNextBirthday from "../lib/functions/birthdays/calcDaysUntilNextBirthday.ts";
+import { has_property } from "../lib/functions/object/hasProperty.ts";
+import { date_to_iso_with_tz } from "../lib/functions/date/timezone.ts";
+import { calc_days_until_next_birthday } from "@/lib/functions/birthdays/calculations.ts";
+import { ISODateFullTZ } from "@/lib/types/date.ts";
+import { Notification } from "./tables/notifications/notifications.ts";
 
 type DexieTable<T extends {id: number}> = Dexie.Table<T, number>;
 
-interface I_Birthbase {
+interface Birthbase {
     birthdays: DexieTable<Birthday>;
     settings: DexieTable<Settings>;
+    notifications: DexieTable<Notification>;
 }
 
 type insertType<T> = T extends DexieTable<infer A> ? A : never;
 
-type getInsert<T extends keyof I_Birthbase> = insertType<I_Birthbase[T]>
+type getInsert<T extends keyof Birthbase> = insertType<Birthbase[T]>
 
-class DexieDB extends Dexie implements I_Birthbase {
+type getSortedBirthdaysOptions = {
+    /**
+     * The Birthdays between the current date and current date + ```maxDayDifference```
+     */
+    maxDayDifference?: number,
+    /**
+     * The max amount of birthdays that will be displayed
+     */
+    maxBirthdays?: number,
+    /**
+     * Gets also the birthdays that are marked as read
+     */
+    all?: boolean,
+    /**
+     * The date or ISOString which is used for comparison. Default value is the current date
+     */
+    comparingDate?: Date|ISODateFullTZ,
+}
+
+// Contants
+enum TABLES {
+    BIRTHDAYS = "birthdays",
+    SETTINGS = "settings",
+    NOTIFICATIONS = "notifications",
+};
+
+enum Errors {
+    NO_FOUND = "Es wurde kein Element gefunden",
+    NO_SUPPORT = ""
+};
+
+const c_db_name = "BirthdayDB";
+
+
+
+// Classes
+class DexieDB extends Dexie implements Birthbase {
     birthdays!: Dexie.Table<Birthday, number>;
     settings!: Dexie.Table<Settings, number>;
+    notifications!: Dexie.Table<Notification, number>;
 
     constructor() {
-        super("BirthdayDB");
+        super(c_db_name);
         this.version(1).stores({
             birthdays: "++id, name, date, marked",
             settings: "id, mode, color, permission.notification, remember",
@@ -33,10 +79,37 @@ class DexieDB extends Dexie implements I_Birthbase {
         this.version(4).stores({
             settings: "++id, mode, color, remember",
         })
+        this.version(5).stores({
+            birthdays: "++id,name,date,marked,remember",
+        })
+        .upgrade(async trans => {
+            return trans.table(TABLES.BIRTHDAYS).toCollection().modify((birthday: Birthday) => {
+                birthday.remember = [];
+            });
+        })
+        this.version(6).stores({
+            birthdays: "++id,name,date,marked,remember",
+        })
+        .upgrade(async trans => {
+            return trans.table(TABLES.BIRTHDAYS).toCollection().modify((birthday: Birthday) => {
+                birthday.date = date_to_iso_with_tz(new Date(birthday.date));
+            });
+        });
+        this.version(7).stores({
+            settings: "++id,name,date,marked,remember,notification",
+        })
+        .upgrade(async trans => {
+            return trans.table(TABLES.SETTINGS).toCollection().modify((settings: Settings) => {
+                settings.notification = false;
+            });
+        });
+        this.version(8).stores({
+            notifications: "++id,timestamp",
+        });
     }
-    create<K extends keyof I_Birthbase, T extends getInsert<K>>(table: K, record: Omit<T, "id">): Promise<T>;
-    create<K extends keyof I_Birthbase, T extends getInsert<K>>(table: K, record: Omit<T, "id">[]): Promise<T[]>;
-    create<K extends keyof I_Birthbase, T extends getInsert<K>>(table: K, record: Omit<T, "id">|Omit<T, "id">[]): Promise<T|T[]> {
+    add<K extends keyof Birthbase, T extends getInsert<K>>(table: K, record: Omit<T, "id">): Promise<T>;
+    add<K extends keyof Birthbase, T extends getInsert<K>>(table: K, record: Omit<T, "id">[]): Promise<T[]>;
+    add<K extends keyof Birthbase, T extends getInsert<K>>(table: K, record: Omit<T, "id">|Omit<T, "id">[]): Promise<T|T[]> {
         return new Promise((resolve, reject) => {
             this.transaction("rw", this[table], async () => {
                 if (Array.isArray(record)) {
@@ -53,7 +126,7 @@ class DexieDB extends Dexie implements I_Birthbase {
                     const _records = await this[table].where("id").anyOf(v).toArray();
                     resolve(_records as T[]);
                 } else {
-                    const _record = await this.read(table, v);
+                    const _record = await this.get(table, v);
                     resolve(_record as T);
                 }
             })
@@ -69,9 +142,9 @@ class DexieDB extends Dexie implements I_Birthbase {
             })
         })
     }
-    read<K extends keyof I_Birthbase, T extends getInsert<K>>(table: K): Promise<T[]>;
-    read<K extends keyof I_Birthbase, T extends getInsert<K>>(table: K, id: number): Promise<T>;
-    read<K extends keyof I_Birthbase, T extends getInsert<K>>(table: K, id?: number): Promise<T|T[]> {
+    get<K extends keyof Birthbase, T extends getInsert<K>>(table: K): Promise<T[]>;
+    get<K extends keyof Birthbase, T extends getInsert<K>>(table: K, id: number): Promise<T>;
+    get<K extends keyof Birthbase, T extends getInsert<K>>(table: K, id?: number): Promise<T|T[]> {
         return new Promise((resolve, reject) => {
             this.transaction("r", (this[table] as Dexie.Table<any, number>), async () => {
                 if (typeof id === "number") {
@@ -92,18 +165,17 @@ class DexieDB extends Dexie implements I_Birthbase {
             })
         })
     }
-    update<K extends keyof I_Birthbase, T extends getInsert<K>>(table: K, record: T): Promise<T>;
-    update<K extends keyof I_Birthbase, T extends getInsert<K>>(table: K, record: T[]): Promise<T[]>;
-    update<K extends keyof I_Birthbase, T extends getInsert<K>>(table: K, record: T|T[]): Promise<T|T[]> {
+    upd<K extends keyof Birthbase, T extends getInsert<K>>(table: K, record: T): Promise<T>;
+    upd<K extends keyof Birthbase, T extends getInsert<K>>(table: K, record: T[]): Promise<T[]>;
+    upd<K extends keyof Birthbase, T extends getInsert<K>>(table: K, record: T|T[]): Promise<T|T[]> {
         return new Promise((resolve, reject) => {
-            let ids: null|number[] = 
-                Array.isArray(record) ? record.map(e => e.id) : null;
             this.transaction("rw", this[table], async () => {
                 if (Array.isArray(record)) {
-                    const a = await (this[table].where("id").anyOf(ids as number[]) as Collection<T>).modify((v, r) => {
+                    const ids = record.map(e => e.id);
+                    const a = await (this[table].where("id").anyOf(ids) as Collection<T>).modify((v, r) => {
                         r.value = record.find((value) => value.id === v.id) as T
                     })
-                    if (a === 0) throw Error("No element was found");
+                    if (a === 0) throw Error(Errors.NO_FOUND);
                 } else {
                     const { id, ..._record } = record;
                     const a = await (this[table] as Dexie.Table<any, number>).update(record.id, _record);
@@ -125,7 +197,7 @@ class DexieDB extends Dexie implements I_Birthbase {
             })
         });
     }
-    Delete<K extends keyof I_Birthbase>(table: K, records: number|number[]): Promise<number> {
+    del<K extends keyof Birthbase>(table: K, records: number|number[]): Promise<number> {
         return new Promise((resolve, reject) => {
             this.transaction("rw", this[table], async () => {
                 if (Array.isArray(records) && records.length === 0) {
@@ -147,50 +219,48 @@ class DexieDB extends Dexie implements I_Birthbase {
     }
 
     /**
-     * Updates the given attributes in the Config
-     * 
-     * If nothing passed attributes be the same
+     * Returns a sorted birthday array (Closest Birthday at first position)
      */
-    async storeSettings (updates: Omit<Partial<Settings>, "id"> = {}): Promise<Settings> {
-        return new Promise(async (resolve, reject) => {
-            try {
-                const settingsArray = await this.read("settings");
-                const settings = await (async () => {
-                    const _settings = getDefaultSettings();
-                    if (settingsArray.length === 0) {
-                        const { id, ...updatedSettings } = ({ ..._settings, ...updates } as Settings);
-                        const newSettings = this.create("settings", updatedSettings);
-                        return newSettings
-                    }
-                    const updatedSettings = { ..._settings, ...settingsArray[0], ...updates }
-                    const newSettings = this.update("settings", updatedSettings);
-                    return newSettings;
-                })();
-                resolve(settings);
-            } catch (e) {
-                console.error()
-                reject(e);
-            }
-        })
-    }
-
-    /**
-     * Returns Birthdays between today and earliest notification
-     * 
-     * @param days earliest notification
-     * @param maxBirthdays How many birthdays will be returned
-     */
-    async getSortedBirthdays (days: number, maxBirthdays: number = 10): Promise<Birthday[]> {
-        const today = new Date();
+    async get_sorted_birthdays (
+        options?: getSortedBirthdaysOptions,
+        // days: number, maxBirthdays: number = 10
+    ): Promise<Birthday[]> {
         return new Promise((resolve, reject) => {
-            if (maxBirthdays === 0) reject("Must be atleast one entry");
+            const today = (() => {
+                if (options) {
+                    if (options.comparingDate) {
+                        if (typeof options.comparingDate === "string") return options.comparingDate;
+                        return date_to_iso_with_tz(options.comparingDate);
+                    } else {
+                        return date_to_iso_with_tz(new Date());
+                    }
+                }
+                return date_to_iso_with_tz(new Date());
+            })();
             this.transaction("r", this.birthdays, async () => {
-                return await this.birthdays.filter((birthday) => {
-                    if (birthday.marked) return false;
-                    const date = new Date(birthday.date);
-                    const difference = (calcDaysUntilNextBirthday(date, today));
-                    return difference <= days;
-                }).limit(maxBirthdays).sortBy("date");
+                let collection = this.birthdays.toCollection();
+                if (options) {
+                    if (has_property(options, "all")) {
+                        if (!options.all) collection = collection.filter((birthday) => !birthday.marked);
+                    }
+                    if (options.maxDayDifference) {
+                        collection = collection.filter((birthday) => {
+                            const difference = (calc_days_until_next_birthday(birthday.date, today));
+                            return difference <= (options.maxDayDifference as number);
+                        })
+                    }
+                    if (options.maxBirthdays) {
+                        collection = collection.limit(options.maxBirthdays);
+                    }
+                }
+                const res = await collection.toArray();
+                return res.sort((a, b) => {
+                    const _a = calc_days_until_next_birthday(a.date, today);
+                    const _b = calc_days_until_next_birthday(b.date, today);
+                    if (_a > _b) return 1;
+                    if (_a < _b) return -1;
+                    return 0;
+                });
             })
             .then((v) => {
                 resolve(v);
@@ -201,7 +271,7 @@ class DexieDB extends Dexie implements I_Birthbase {
         })
     }
 
-    getStorageSize = async (): Promise<number | undefined> => {
+    get_storage_size = async (): Promise<number | undefined> => {
         return new Promise(async (resolve, reject) => {
             if (navigator.storage) {
                 const estimation = await navigator.storage.estimate();
@@ -218,11 +288,4 @@ class DexieDB extends Dexie implements I_Birthbase {
 
 const db = new DexieDB();
 
-/**
- * Settings Config at the start of the application
- * 
- * If App has no Config, Config will be filled with undefined values
- */
-const __INI_APP_SETTINGS__ = await db.storeSettings();
-
-export { DexieDB, db, __INI_APP_SETTINGS__};
+export { DexieDB, db, TABLES };
