@@ -1,27 +1,19 @@
 import { defineConfig, loadEnv } from 'vite'
 import path from "path"
 import react from '@vitejs/plugin-react'
+import tailwindcss from "@tailwindcss/vite";
 
 const host = process.env.TAURI_ENV_HOST;
 const mobile = process.env.TAURI_ENV_PLATFORM;
 
-console.log(mobile);
-console.log(process.env.TAURI_DEV_HOST);
-console.log(process.env.TAURI_ENV_HOST);
-
-console.log(process.argv);
-
 // https://vitejs.dev/config/
 export default defineConfig(({ command, mode, isPreview }) => {
-  console.log("Mode:", mode);
-  console.log(command);
-  console.log("Preview:", isPreview);
-
-  console.log("Env:", loadEnv(mode, process.cwd()));
-
   return {
     publicDir: "public",
-    plugins: [react()],
+    plugins: [
+      react(),
+      tailwindcss(),
+    ],
     resolve: {
       alias: {
         "@": path.resolve(__dirname, "./src"),
@@ -36,6 +28,7 @@ export default defineConfig(({ command, mode, isPreview }) => {
   //   // 2. tauri expects a fixed port, fail if that port is not available
     server: {
       host: mobile === "android" ? '0.0.0.0' : (host || false),
+      // host: mobile === (host || false),
       port: 1420,
       strictPort: true,
   //     host: mobile ? "0.0.0.0" : false,
@@ -51,10 +44,23 @@ export default defineConfig(({ command, mode, isPreview }) => {
         ignored: ["**/src-tauri/**"],
       },
     },
+    optimizeDeps: {
+      esbuildOptions: {
+        target: "esnext"
+      }
+    },
     build: {
-      target: ["esNext"],
+      rollupOptions: {
+        output: {
+          format: "es"
+        }
+      },
+      target: "esnext",
       minify: !process.env.TAURI_DEBUG ? "esbuild" : false,
       sourcemap: true/*!!process.env.TAURI_DEBUG*/,
+    },
+    worker: {
+      format: "es",
     }
   }
 })
